+++ FOX/functions/rdf.py
@@ -69,8 +69,7 @@ def get_rdf(dist: np.ndarray,
 
     dist_shape = dist.shape
     dens_mean = dist_shape[2] / ((4/3) * np.pi * (0.5 * dist.max(axis=(1, 2)))**3)
-    dist2 = dist / dr
-    dist2 = dist2.astype(np.int32, copy=False)
+    dist2 = (dist / dr).astype(np.int32, copy=False)
     dist2.shape = dist_shape[0], dist_shape[1] * dist_shape[2]
 
     dens = np.array([np.bincount(i, minlength=idx_max)[:idx_max] for i in dist2], dtype=float)