+++ CHANGELOG.rst
@@ -5,6 +5,14 @@ Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to `Semantic Versioning <http://semver.org/>`_.
 
+0.6.1
+*****
+* Added an additionl memory consumption level to `MultiMolecule.init_rdf()`.
+* Ensure that the 'constraints' column is always present in the ARMC parameter DataFrame.
+* `_xyz_to_hdf5()` no longer crashes when `mol_list=None`.
+* Switched the `AssertionLib` package from GitHub to PyPi.
+
+
 0.6.0
 *****
 * Many minor (consistancy) changes and codestyle improvements.